+++ packages/angular/index.js
@@ -1,0 +1,7 @@
+export const adapter = {
+  framework: "angular",
+  packages: [
+    "@angular-eslint/eslint-plugin-template",
+    "@angular-eslint/template-parser"
+  ]
+};