+++ packages/vue/index.js
@@ -1,0 +1,4 @@
+export const adapter = {
+  framework: "vue",
+  packages: ["eslint-plugin-vue"]
+};